+++ packages/nextjs/components/signup-component/pages/JoinWithUs.tsx
@@ -1,23 +1,13 @@
 import React from "react";
-import BackButton from "../BackButton";
-import Button from "../Button";
-import Typography from "~~/components/Typography";
+import Typography from "~~/components/signup-component/Typography";
 
 interface JoinWithUsProps {
   setTab: React.Dispatch<React.SetStateAction<number>>;
 }
 
-const JoinWithUs: React.FC<JoinWithUsProps> = ({ setTab }) => {
-  const handleNext = () => {
-    setTab(1);
-  };
-  const handleBack = () => {
-    window.history.back();
-  };
-
+const JoinWithUs: React.FC<JoinWithUsProps> = () => {
   return (
     <>
-      <BackButton onClick={handleBack} />
       <div className="justify-center flex flex-col h-screen items-center gap-10">
         <Typography variant="title">Join With Us</Typography>
         <Typography variant="desc">
@@ -25,14 +15,6 @@ const JoinWithUs: React.FC<JoinWithUsProps> = ({ setTab }) => {
           in their style. If you already have a character created with Ceptor Club, select the Import button and we will
           fill out the questions for you!
         </Typography>
-        <section className="grid grid-cols-1 md:grid-cols-2 gap-4 w-full text-center md:w-[824px]">
-          <Button variant="fill" onClick={handleNext}>
-            Sign Up
-          </Button>
-          <Button variant="outline" onClick={handleBack}>
-            Back
-          </Button>
-        </section>
       </div>
     </>
   );
